refactor: freeform and welcome post modal/page support (#1835)
* refactor: post content applicable to freeform post

* feat: edit post option

* fix: nullable
